backend/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
flask==2.3.3